moved hooks out of directory, ci seems to fail because for whatever reason
